Moving gl utils to common folder, along with debugview now using advanced OBJ loading
