+++ gldemo/demo_data/headset.mtl
@@ -1,0 +1,24 @@
+# Blender MTL File: 'headset.blend'
+# Material Count: 2
+
+newmtl Headset
+Ns 0.000000
+Ka 1.000000 1.000000 1.000000
+Kd 0.123710 0.123710 0.123710
+Ks 0.000000 0.000000 0.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.450000
+d 1.000000
+illum 1
+map_Kd HeadsetBake.png
+
+newmtl Logo
+Ns 0.000000
+Ka 1.000000 1.000000 1.000000
+Kd 0.800000 0.800000 0.800000
+Ks 0.000000 0.000000 0.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.450000
+d 1.000000
+illum 1
+map_Kd HeadsetBake.png
